# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DSA Studio API
  description: REST API for the DSA Studio learning platform — Phase 1 Foundation endpoints.
  version: 0.1.0
  contact:
    name: DSA Studio Team

servers:
  - url: http://localhost:4000/api
    description: Local development

tags:
  - name: Health
  - name: Auth
  - name: Topics
  - name: Questions

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object

    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        learningLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        dailyTarget:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        totalQuestionsSolved:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Topic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        difficulty:
          type: string
          enum: [basic, intermediate, advanced]
        orderIndex:
          type: integer
        icon:
          type: string
          nullable: true
        prerequisiteIds:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        totalQuestions:
          type: integer

    TestCase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        input:
          type: string
        expectedOutput:
          type: string
        explanation:
          type: string
          nullable: true
        isSample:
          type: boolean
        isHidden:
          type: boolean
        orderIndex:
          type: integer

    QuestionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        topicId:
          type: string
          format: uuid
        topicSlug:
          type: string
          nullable: true
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [basic, intermediate, advanced]
        hints:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    QuestionDetail:
      allOf:
        - $ref: '#/components/schemas/QuestionSummary'
        - type: object
          properties:
            constraints:
              type: string
              nullable: true
            inputFormat:
              type: string
              nullable: true
            outputFormat:
              type: string
              nullable: true
            timeComplexity:
              type: string
              nullable: true
            spaceComplexity:
              type: string
              nullable: true
            testCases:
              type: array
              items:
                $ref: '#/components/schemas/TestCase'

    PaginatedMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

paths:
  /health:
    get:
      tags: [Health]
      summary: API health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          status:
                            type: string
                            example: ok
                          service:
                            type: string
                          version:
                            type: string
                          timestamp:
                            type: string
                            format: date-time

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Email or username already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (client discards token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
        '401':
          description: Unauthorized

  /topics:
    get:
      tags: [Topics]
      summary: List topics
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [basic, intermediate, advanced]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated topic list

  /topics/progress:
    get:
      tags: [Topics]
      summary: User completion per topic
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Topic progress breakdown

  /topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topic detail
        '404':
          description: Topic not found

  /topics/{id}/questions:
    get:
      tags: [Topics]
      summary: List questions for a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questions in topic

  /questions:
    get:
      tags: [Questions]
      summary: List questions with filters
      parameters:
        - name: topicId
          in: query
          schema:
            type: string
            format: uuid
        - name: topicSlug
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [basic, intermediate, advanced]
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated question list

  /questions/{id}:
    get:
      tags: [Questions]
      summary: Get question detail (sample test cases only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Question detail
        '404':
          description: Question not found